+++ src/js/api/index.js
@@ -1,5 +1,6 @@
 import globalAPI from './global';
-import internalAPI from './internal';
+import hooksAPI from './hooks';
+import stateAPI from './state';
 import instanceAPI from './instance';
 import componentAPI from './component';
 import * as util from '../util/index';
@@ -16,7 +17,8 @@ UIkit.instances = {};
 UIkit.elements = [];
 
 globalAPI(UIkit);
-internalAPI(UIkit);
+hooksAPI(UIkit);
+stateAPI(UIkit);
 instanceAPI(UIkit);
 componentAPI(UIkit);