+++ src/js/core/scrollspy-nav.js
@@ -20,15 +20,27 @@ export default function (UIkit) {
             offset: 0
         },
 
+        computed: {
+
+            links() {
+                return this.$el.find('a[href^="#"]').filter((i, el) => el.hash);
+            },
+
+            elements() {
+                return this.closest ? this.links.closest(this.closest) : this.links;
+            },
+
+            targets() {
+                return $(this.links.toArray().map(el => el.hash).join(','));
+            }
+
+        },
+
         update: [
 
             {
 
                 read() {
-                    this.links = this.$el.find('a[href^="#"]').filter((i, el) => el.hash);
-                    this.elements = (this.closest ? this.links.closest(this.closest) : this.links);
-                    this.targets = $($.map(this.links, (el) => el.hash).join(','));
-
                     if (this.scroll) {
                         this.links.each((_, el) => UIkit.scroll(el, {offset: this.offset || 0}));
                     }